chore: move MdxPost to components level
