+++ apps/website/src/components/MdxPost/components/FeaturedImage/FeaturedImage.tsx
