+++ apps/website/src/components/MdxPost/components/FeaturedImage/index.ts
